*: expose command line arguments
The command line arguments are going to need to be included in a
provenance attestation.
