+++ src/lib.rs
@@ -15,6 +15,8 @@ pub mod in_toto;
 pub mod nsm;
 
 use hyper::body::Bytes;
+use std::fmt;
+use std::net::{IpAddr, Ipv4Addr};
 
 pub struct Artifact {
     pub name: String,
@@ -35,5 +37,38 @@ impl ArtifactFormat {
             DockerArchive => "docker-archive",
             OciArchive => "oci-archive",
         }
+    }
+}
+
+#[derive(Clone, Copy, clap::Parser)]
+#[command(version)]
+pub struct Config {
+    #[arg(default_value_t = IpAddr::V4(Ipv4Addr::LOCALHOST), long, short)]
+    pub address: IpAddr,
+
+    #[arg(default_value_t = 8080, long, short)]
+    pub port: u16,
+
+    #[arg(default_value_t = SpdxGenerator::SyftBinary, long, short)]
+    pub spdx: SpdxGenerator,
+
+    #[arg(long, short)]
+    pub one_shot: bool,
+}
+
+#[derive(Clone, Copy, clap::ValueEnum)]
+pub enum SpdxGenerator {
+    SyftBinary,
+    SyftDockerContainer,
+}
+
+impl fmt::Display for SpdxGenerator {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        use clap::ValueEnum;
+
+        self.to_possible_value()
+            .expect("no skipped variants")
+            .get_name()
+            .fmt(f)
     }
 }
